requirements.txt
asgiref==3.2.3
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
cffi==1.13.2
chardet==3.0.4
colorama==0.4.3
cryptography==2.8
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.3
django-allauth==0.39.1
django-cors-headers==3.0.2
django-crispy-forms==1.6.0
django-filter==2.2.0
django-heroku==0.3.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.9.4
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
Markdown==3.1.1
markdown2==2.3.1
mccabe==0.6.1
oauthlib==3.0.1
pep8==1.7.1
psycopg2==2.8.4
pycodestyle==2.5.0
pycparser==2.19
pylint==2.3.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.7.9
whitenoise==5.0.1
wrapt==1.11.2
